- Done: sending push notif to all and to one user from api - ToDo: get and register PlayerID
